modularized SiloNode
