+++ combineNodes.js
@@ -1,141 +1,46 @@
 // import state class for instanceof check
 const StateNode = require('./stateNode.js');
+const SiloNode = require('./SiloNode.js');
 
 // ==================> SILO TESTING <=================== \\
 
-const AppState = new StateNode('AppState');
-// AppState.name = 'AppState'; -> optional if not set in constructor
+// const AppState = new StateNode('AppState');
+// // AppState.name = 'AppState'; -> optional if not set in constructor
 
-AppState.initializeState({
-  name: 'Han',
-  age: 25,
-  cart: {one:1, two:2}
-})
+// AppState.initializeState({
+//   name: 'Han',
+//   age: 25,
+//   cart: {one:1, two:2}
+// })
 
-AppState.initializeModifiers({
-  age: {
-    incrementAge: (current, payload) => {
-      return current + payload;
-    }
-  },
-  cart: {
-    increment: (current, index, payload) => {
-      return ++current;
-    }
-  }
-});
+// AppState.initializeModifiers({
+//   age: {
+//     incrementAge: (current, payload) => {
+//       return current + payload;
+//     }
+//   },
+//   cart: {
+//     increment: (current, index, payload) => {
+//       return ++current;
+//     }
+//   }
+// });
 
-const NavState = new StateNode('NavState');
-NavState.parent = 'AppState';
+// const NavState = new StateNode('NavState');
+// NavState.parent = 'AppState';
 
-NavState.initializeState({
-  nav: 'Nav'
-})
+// NavState.initializeState({
+//   nav: 'Nav'
+// })
 
-const ButtState = new StateNode('ButtState');
-ButtState.parent = 'NavState';
+// const ButtState = new StateNode('ButtState');
+// ButtState.parent = 'NavState';
 
-ButtState.initializeState({
-  butt: 'Butt'
-})
+// ButtState.initializeState({
+//   butt: 'Butt'
+// })
 
 //==================> SILO TESTING ENDED <===================\\
-
-class siloNode {
-  constructor(val, parent = null, modifiers = {}, isAnObject = false) {
-    this._value = val;
-    this._modifiers = modifiers;
-    this._queue = [];
-    this._subscribers = [];
-    this._parent = parent; // circular silo node
-    this._isAnObject = isAnObject; // controls the type of middleware
-
-    // bind
-    this.linkModifiers = this.linkModifiers.bind(this);
-    this.runModifiers = this.runModifiers.bind(this);
-
-    // invoke functions
-    this.linkModifiers(this.modifiers);
-    this.runQueue = this.runModifiers();
-  }
-
-  get value() {
-    return this._value;
-  }
-
-  get modifiers() {
-    return this._modifiers;
-  }
-
-  get queue() {
-    return this._queue;
-  }
-
-  get parent() {
-    return this._parent;
-  }
-
-  get isAnObject() {
-    return this._isAnObject;
-  }
-
-  runModifiers() {
-    let running = false; // prevents multiple calls from being made if already running
-
-    async function run() {
-      if (running === false) { // prevents multiple calls from being made if already running
-        running = true;
-  
-        while (this.queue.length > 0) {
-          this.value = await this.queue.shift()();
-          // tell subscribers!!!
-          console.log("in while loop", this.value); // test purposes only
-        }              
-      } else {
-        return 'in progress...';
-      }
-    }
-    return run;
-  };
-
-  linkModifiers(stateModifiers) {
-    if (!stateModifiers) return;
-    const that = this;
-
-    // looping through every modifier added by the dev
-    Object.keys(stateModifiers).forEach(modifierKey => {
-      const modifier = stateModifiers[modifierKey];
-
-      if (typeof modifier !== 'function' ) throw new TypeError(); 
-      // adds middleware that will affect the value of this node
-      else if (!this.isAnObject) {
-        // wrap the dev's modifier function so we can pass the current node value into it
-        const linkedModifier = async (payload) => await modifier(that.value, payload); 
-
-        // the function that will be called when the dev tries to call their modifier
-        stateModifiers[modifierKey] = payload => {
-          // wrap the linkedModifier again so that it can be added to the async queue without being invoked
-          const callback = async () => await linkedModifier(payload);
-          that.queue.push(callback);
-          that.runQueue();
-        }
-      }
-      // adds middleware that will affect the value of a child node
-      else if (this.isAnObject) {
-        // wrap the dev's modifier function so we can pass the current node value into it
-        const linkedModifier = async (index, payload) => await modifier(that.value[index].value, index, payload); 
-
-        // the function that will be called when the dev tries to call their modifier
-        stateModifiers[modifierKey] = (index, payload) => {
-          // wrap the linkedModifier again so that it can be added to the async queue without being invoked
-          const callback = async () => await linkedModifier(index, payload);
-          that.value[index].queue.push(callback);
-          that.value[index].runQueue();
-        }
-      }
-    })
-  }
-}
 
 // ===========> async TEST stuff <========== \\
 
@@ -145,7 +50,7 @@ class siloNode {
 //   })
 // }
 
-// const nodeA = new siloNode(5, null, {
+// const nodeA = new SiloNode(5, null, {
 //   increment: (current, payload) => {
 //     return current + payload;
 //   },
@@ -176,27 +81,27 @@ const silo = {};
 // IMPORTANT nested object nodes are named after their parent and the key: ex: cart_one
 function handleNestedObject(objName, obj, parent) {
   const objChildren = {};
-  const node = new siloNode(objChildren, parent, obj.modifiers, true);   // the true argument indicates that this is a parent object node
+  const node = new SiloNode(objChildren, parent, obj.modifiers, true);   // the true argument indicates that this is a parent object node
   const keys = Array.isArray(obj.value) ? obj : Object.keys(obj.value);
   
   if (Array.isArray(obj.value) && obj.value.length > 0) {
     obj.value.forEach((val, i) => {
       if (typeof val === 'object') objChildren[`${objName}_${i}`] = handleNestedObject(`${objName}_${i}`, {value: val}, node);
-      else objChildren[`${objName}_${i}`] = new siloNode(val, node);
+      else objChildren[`${objName}_${i}`] = new SiloNode(val, node);
     })
   } 
   
   else if (keys.length > 0) {
     keys.forEach(key => {
       if (typeof obj.value[key] === 'object') objChildren[`${objName}_${key}`] = handleNestedObject(key, {value: obj.value[key]}, node);
-      else objChildren[`${objName}_${key}`] = new siloNode(obj.value[key], node);
+      else objChildren[`${objName}_${key}`] = new SiloNode(obj.value[key], node);
     })
   }
 
   return node;
 }
 
-// combineNodes takes all of the stateNodes created by the developer. It then creates siloNodes from the
+// combineNodes takes all of the stateNodes created by the developer. It then creates SiloNodes from the
 // stateNodes and organizes them into a single nested object, the silo
 
 combineNodes = (...args) => {
@@ -240,19 +145,19 @@ combineNodes = (...args) => {
     hashTable[nodeName].forEach(child => {
 
       const nodeVal = {};
-      allChildren[child.name] = new siloNode(nodeVal, parent);
+      allChildren[child.name] = new SiloNode(nodeVal, parent);
       const thisStateNode = child;
       const thisSiloNode = allChildren[child.name];
       const stateObj = child.state;
 
-      // create siloNodes for all the variables in the child state node
+      // create SiloNodes for all the variables in the child state node
       Object.keys(stateObj).forEach(varName => {
         // handles non primitive data types
         if (typeof stateObj[varName].value === 'object') {
           nodeVal[varName] = handleNestedObject(varName, stateObj[varName], thisSiloNode);
         }
         // primitives only
-        else nodeVal[varName] = new siloNode(stateObj[varName].value, thisSiloNode, stateObj[varName].modifiers);
+        else nodeVal[varName] = new SiloNode(stateObj[varName].value, thisSiloNode, stateObj[varName].modifiers);
       })
 
       // recurse for grandbabiessss
@@ -275,7 +180,7 @@ combineNodes = (...args) => {
   });
 }
 
-combineNodes(ButtState, NavState, AppState); // testing purposes
+// combineNodes(ButtState, NavState, AppState); // testing purposes
 
 // ==========> TESTS that calling a parent function will modify its child for nested objects <========== \\
